base url w/o trailing /
